refactor: rename route /tests -> /subjects
