+++ src/components/routes/layouts/root/MobileNav/MobileNav.tsx
@@ -13,7 +13,7 @@ export default async function MobileNav() {
   const session = await auth();
   const navItems: NavItem[] = [
     {
-      href: "/tests",
+      href: "/subjects",
       label: "Тесты",
       icon: <ListTodoIcon />,
     },
