+++ src/components/routes/my/subjects/SubjectsList.tsx
@@ -50,7 +50,7 @@ export default function SubjectsList() {
       <ul>
         {subjects?.map(({ id, title }) => (
           <li key={id} className="flex justify-between">
-            <Link className="w-full" href={`/tests/${id}`}>
+            <Link className="w-full" href={`/subjects/${id}`}>
               <Button className="w-full justify-start" variant="ghost">
                 {title}
               </Button>
